#8 Renamed package and artifact
