+++ pom.xml
@@ -1,7 +1,7 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <modelVersion>4.0.0</modelVersion>
-  <groupId>com.github.glusk2</groupId>
+  <groupId>com.github.glusk</groupId>
   <artifactId>tecajnica</artifactId>
   <version>1.0-SNAPSHOT</version>
   <name>tecajnica</name>
