+++ src/main/java/com/github/glusk/App.java
@@ -1,4 +1,4 @@
-package com.github.glusk2;
+package com.github.glusk;
 
 import java.net.URL;
 import java.time.LocalDate;
